- refactor clip into video - implement getVideo API
