+++ packages/frontend/components/Video/VideoCarousel.tsx
@@ -1,18 +1,12 @@
 import React, { useState, useEffect } from 'react'
-import { VideoInterface } from './Video'
+import { VideoInterface } from '../../interfaces/video'
 import VideoPlayer from './VideoPlayer'
 import styles from './videoBoard.module.scss'
 import Carousel from 'react-multi-carousel'
-import { getVideo } from '../../api/temp-video'
+import { getVideo } from '../../api/video'
 import 'react-multi-carousel/lib/styles.css'
 
 const responsive = {
-  superLargeDesktop: {
-    // the naming can be any, depends on you.
-    breakpoint: { max: 4000, min: 3000 },
-    items: 1,
-    partialVisibilityGutter: 40,
-  },
   desktop: {
     breakpoint: { max: 3000, min: 1024 },
     items: 1,
@@ -20,10 +14,6 @@ const responsive = {
   },
   tablet: {
     breakpoint: { max: 1024, min: 464 },
-    items: 1,
-  },
-  mobile: {
-    breakpoint: { max: 464, min: 0 },
     items: 1,
   },
 }
@@ -37,14 +27,14 @@ export default function VideoCarousel() {
       setVideos(videos)
     }
     onStart()
-  })
+  }, [])
 
   return (
     <span className={styles.carousel}>
       {videos && (
         <Carousel responsive={responsive} centerMode infinite>
-          {videos.map((video) => (
-            <VideoPlayer key={video.id} video={video} />
+          {videos.map((video, i) => (
+            <VideoPlayer key={i} video={video} />
           ))}
         </Carousel>
       )}
