--- packages/frontend/interfaces/clip.ts
@@ -1,4 +1,0 @@
-export interface ClipInterface {
-  creator: string
-  videoEmbedUrl: string
-}
