+++ packages/frontend/interfaces/video.ts
@@ -1,0 +1,5 @@
+export interface VideoInterface {
+  creator: string
+  videoEmbedUrl: string
+  title?: string
+}
